asr_web_app/requirements.txt
Flask
Flask-SocketIO
websocket-client
python-engineio
python-socketio
python-multipart
eventlet
ngrok-api
requests
aiortc
numpy
aiohttp
av 